feat(nvmeadm): support of subsystem state synchronisation
NVMe Subsystem represents an in-memory snapshot of an NVMe subsystem
that exists on the system. When state of the underlying subsystem changes,
there is no possibility to get this change reflected in existing Subsystem object
other than recreate it from scratch.
This change introduces a new function to allow explicit synchronisation on demand
for existing Subsystem objects.

Signed-off-by: Mikhail Tcymbaliuk <mtzaurus@gmail.com>