fix: use cursor to read from byte buffer (#78)
Previous ByteBuffer implementation was incorrect because the
slice was created on demand, thus meaning it did not move..
As such the contents from the start were always returned.

Adds a test to make sure we at least don't get stuck though
a better test would be better to ensure we read the correct data.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>