refactor(events): add default replicas and allow parsing (#126)
Adds a default number of replicas to the bus itself.
Makes replicas part of the event handle api, allowing publishers
to set it.
Change the subscriber to only get the stream and error out if
not created within a given timeout - this ensures that the
subscriber doesn't mistakenly set the stream's n of replicas.

We should also consider if we are doing the right thing here,
as it seems that allowing any publisher to create the stream
doesn't allow for easy updates, as they all share the same
stream. Perhaps we should have a single component responsible
for setting up the stream?

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>