build: try to fetch parent root dir
When the release.sh is sourced from another repo, the parent root may not
be a dir away from the script dirname.
Instead we figure out by trying well known env variables and even using
the git top-level.

Adds a debug-slim for easy quick builds in debug mode for testing.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>